Moved the DeskClientBuilder to own class.
