+++ api-client/src/test/java/com/desk/java/apiclient/util/DeskClientUtilsTest.java
@@ -28,10 +28,10 @@ package com.desk.java.apiclient.util;
 
 import org.junit.Test;
 
-import static com.desk.java.apiclient.DeskClient.OAUTH_ACCESS_URL;
-import static com.desk.java.apiclient.DeskClient.OAUTH_AUTHORIZE_URL;
-import static com.desk.java.apiclient.DeskClient.OAUTH_REQUEST_URL;
-import static com.desk.java.apiclient.DeskClient.PROTOCOL_CONNECT;
+import static com.desk.java.apiclient.DeskClientBuilder.OAUTH_ACCESS_URL;
+import static com.desk.java.apiclient.DeskClientBuilder.OAUTH_AUTHORIZE_URL;
+import static com.desk.java.apiclient.DeskClientBuilder.OAUTH_REQUEST_URL;
+import static com.desk.java.apiclient.DeskClientBuilder.PROTOCOL_CONNECT;
 import static org.junit.Assert.assertEquals;
 
 /**
