+++ rxjava-api-client/src/main/java/com/desk/java/apiclient/service/RxInboundMailboxService.java
@@ -26,7 +26,7 @@
 
 package com.desk.java.apiclient.service;
 
-import com.desk.java.apiclient.RxDeskClient;
+import com.desk.java.apiclient.DeskClientBuilder;
 import com.desk.java.apiclient.model.ApiResponse;
 import com.desk.java.apiclient.model.InboundMailbox;
 
@@ -39,7 +39,7 @@ import static com.desk.java.apiclient.service.InboundMailboxService.INBOUND_MAIL
 /**
  * <p>
  * Service interfacing with the Desk Inbound Mailboxes endpoint.
- * This service supports {@link RxDeskClient.AuthType#API_TOKEN} authentication.
+ * This service supports {@link DeskClientBuilder.AuthType#API_TOKEN} authentication.
  * </p>
  * <p>
  * Created by Matt Kranzler on 6/26/15.
