+++ rxjava-api-client/src/main/java/com/desk/java/apiclient/service/RxTopicService.java
@@ -26,7 +26,7 @@
 
 package com.desk.java.apiclient.service;
 
-import com.desk.java.apiclient.RxDeskClient;
+import com.desk.java.apiclient.DeskClientBuilder;
 import com.desk.java.apiclient.model.ApiResponse;
 import com.desk.java.apiclient.model.Article;
 import com.desk.java.apiclient.model.SortDirection;
@@ -45,7 +45,7 @@ import static com.desk.java.apiclient.service.TopicService.TOPICS_URI;
 /**
  * <p>
  * Service interfacing with the Desk Topics endpoint. This service supports
- * {@link RxDeskClient.AuthType#API_TOKEN} authentication.
+ * {@link DeskClientBuilder.AuthType#API_TOKEN} authentication.
  * </p>
  * <p>
  * Created by Matt Kranzler on 6/19/15.
